Combine dependency injection registrations
Previously, there were two places where the dependencies were registered. This consolidates it to a single entry point that is easier to follow and ensures the services are available where expected.
